restructuring mostly
