+++ src/builtin/echo.c
