+++ src/executor/exec_pipelines.c
@@ -12,7 +12,7 @@
 
 #include "executor.h"
 
-int	spawn_and_run(t_node *node, t_shell *shell, int count, int *prev_fd)
+int	spawn_and_run(t_node *node, t_state *shell, int count, int *prev_fd)
 {
 	pid_t		child_pid;
 	
@@ -22,11 +22,11 @@ int	spawn_and_run(t_node *node, t_shell *shell, int count, int *prev_fd)
 		if (create_pipe(node, shell, prev_fd)
 			return (ERROR_PIPELINE);
 	}
-	if (child_pid != 0) //except builtins shouldn't be forked: unset and env need ability to permanently modify parent ENV variables
+	if (child_pid != 0 && node->cmd->builtin == false) //builtins shouldn't be forked: unset and env need to modify parent ENV vars for good
 	{
 		if (fork_child(&child_pid, shell))
 			return (ERROR_FORKING);
-		shell->pids[count] = child_pid;
+		shell->pids[count] = child_pid; //this creates a copy to child i think.
 	}
 	if (child_pid == 0)
 	{
@@ -34,7 +34,8 @@ int	spawn_and_run(t_node *node, t_shell *shell, int count, int *prev_fd)
 			return (ERROR_REDIR);
 		run_node(node, shell, count, *prev_fd);
 	}
-	close_parent_pps(node, shell);
+	if (close_parent_pps(node, shell))
+		return (ERROR_CLOSING_FD);
 	return (SUCCESS);
 }
 
@@ -42,20 +43,14 @@ int	fork_child(pid_t *child_pid, t_state *shell);
 {
 	*child_pid = fork();
 	if (*child_pid == -1)
-	{
-		shell->exit_status = ERROR_FORKING;
 		return (ERROR_FORKING);
-	}
 	return (SUCCESS);
 }
 
 int	create_pipe(t_node *node, t_state *shell, int *prev_fd)
 {
 	if (pipe(node->pipe_fds))
-	{
-		shell->exit_status = ERROR_PIPELINE;
 		return (ERROR_PIPELINE);
-	}
 	*prev_fd = node->pipe_fds[1];
 	return (SUCCESS);
 }
@@ -64,33 +59,34 @@ int	redirections(t_node *node, t_state *shell, int prev_fd)
 {
 	if (node->prev)
 	{
-		dup2(prev_fd, STDIN_FILENO);
+		if (dup2(prev_fd, STDIN_FILENO)
+	  		return (ERROR_REDIR);
 		if (close(prev_fd))
-			shell->exit_status = ERROR_REDIR;
+			return (ERROR_CLOSING_FD);
 	}
 	else if (node->prev == NULL)
 	{
 		if (close(prev_fd))
-			shell->exit_status = ERROR_REDIR;
+			return (ERROR_CLOSING_FD);
 	}
 	if (node->next)
 	{
-		dup2(node->pipe_fds[1], STDOUT_FILENO);
+		if (dup2(node->pipe_fds[1], STDOUT_FILENO))
+			return (ERROR_REDIR);
 		if (close(node->pipe_fds[0]))
-			shell->exit_status = ERROR_REDIR;
+			return (ERROR_CLOSING_FD);
 	}
-	return (shell->exit_status);
-
+	return (SUCCESS);
 }
 
 int	close_parent_pps(t_node *node, t_state *shell)
 {
 	if (close(node->pipe_fds[0]))
-		shell->exit_status = ERROR_REDIR;
+		return (ERROR_CLOSING_FD);
 	if (node->prev)
 	{
 		if (close(node->prev->pipe_fds[1]))
-			shell->exit_status = ERROR_REDIR;
+			return (ERROR_CLOSING_FD);
 	}
 	return (SUCCESS);
 }
